requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

librosa>=0.10.1
numpy>=1.24.0
pydub>=0.25.1
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
soundfile>=0.12.1

safetensors
huggingface-hub#for interfrence with Hugging Face models